* big news: all PythonApi method signatures are now determined from gccxml output * consequent big news: I have been rudely forced to get over my inability to distinguish between size_t and ssize_t, which has caused a fair number of knock-on changes in PythonMapper_whatever files, and associated tests. generating structs will present similar challenges. still, all for the best.
various interesting details follow

* renamed PythonMapper.ForceCleanup to DemandCleanup, to reflect the oft-futile nature of GC requests
* function specs in _dispatcher_methods have been corrected from 'size' to 'ssize' where appropriate (we should be autogenerating these too, now)
* _extra_dgttypes is gone, because we pick up the required dgts automatically now
* SLICE_HACK_TEMPLATE_TEMPLATE is now gone (size/ssize...)
* _mgd_api_functions now contains only names of those API functions for which we want a managed implementation
* _mgd_function_prototypes now contains C prototypes for all 'IC_" functions, rather than having some of them represented by name/dgt_spec pairs (as were in the old _mgd_api_functions)
* platform.py is gone; effectively replaced by type_codes.py (which contains one platform-specific dict)
* free functions in apiwrangler moved to new dispatchergen.py, magicmethodgen.py, utils_codegen.py, and preexisting utils.py
* horrid horrid utils_c.py is gone
* apiwrangler operation overall made cleaner/clearer

by: William

+++ src/PythonMapper_list.cs
@@ -80,7 +80,7 @@ namespace Ironclad
         }
 
         public override IntPtr 
-        PyList_New(uint length)
+        PyList_New(int length)
         {
             if (length == 0)
             {
@@ -90,11 +90,11 @@ namespace Ironclad
             PyListObject list = new PyListObject();
             list.ob_refcnt = 1;
             list.ob_type = this.PyList_Type;
-            list.ob_size = length;
-            list.allocated = length;
+            list.ob_size = (uint)length;
+            list.allocated = (uint)length;
             
-            uint bytes = length * CPyMarshal.PtrSize;
-            list.ob_item = this.allocator.Alloc(bytes);
+            int bytes = length * CPyMarshal.PtrSize;
+            list.ob_item = this.allocator.Alloc((uint)bytes);
             CPyMarshal.Zero(list.ob_item, bytes);
             
             IntPtr listPtr = this.allocator.Alloc((uint)Marshal.SizeOf(typeof(PyListObject)));
@@ -154,7 +154,7 @@ namespace Ironclad
         
         
         public override int
-        PyList_SetItem(IntPtr listPtr, uint index, IntPtr itemPtr)
+        PyList_SetItem(IntPtr listPtr, int index, IntPtr itemPtr)
         {
             if (!this.HasPtr(listPtr))
             {
@@ -195,12 +195,12 @@ namespace Ironclad
         
         
         public override IntPtr
-        PyList_GetSlice(IntPtr listPtr, uint start, uint stop)
+        PyList_GetSlice(IntPtr listPtr, int start, int stop)
         {
             try
             {
                 List list = (List)this.Retrieve(listPtr);
-                List sliced = (List)list[new Slice((int)start, (int)stop)];
+                List sliced = (List)list[new Slice(start, stop)];
                 return this.Store(sliced);
             }
             catch (Exception e)
@@ -212,7 +212,7 @@ namespace Ironclad
 
 
         public override IntPtr
-        PyList_GetItem(IntPtr listPtr, uint idx)
+        PyList_GetItem(IntPtr listPtr, int idx)
         {
             try
             {
